Uploaded files as per Kostiantyns request
I simply copied over the relevant files for what you asked for so I don't think the paths are correct for an execution of the app.py to yield any result. However you may now have an overview of how the process looks. 

Disregard the (final) at the end of the file names, it is not final...